# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "endPoind ",
        "version": "1.0"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/liste_formation": {
            "get": {
                "tags": [
                    "Formations"
                ],
                "summary": "Liste des formations",
                "description": "Récupère une liste de toutes les formations.",
                "operationId": "cc1fd3a9bd1eaeb7fa338436d0841c09",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "/components/schemas/Formation"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        },
        "/api/create_formation": {
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Créer une nouvelle formation",
                "description": "Crée une nouvelle formation.",
                "operationId": "f1f81bac3e60cc23a30b8a064c80efdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_formation": {
                                        "type": "string"
                                    },
                                    "dure_formation": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation créée avec succès"
                    },
                    "500": {
                        "description": "Formation non créer"
                    }
                }
            }
        },
        "/api/exister_formation/{formation}": {
            "get": {
                "tags": [
                    "Formations"
                ],
                "summary": "Afficher une formation spécifique",
                "description": "Affiche les détails d'une formation spécifique.",
                "operationId": "9efcb3b4890299dd7d9e17f21fa0d70a",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID de la formation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/components/schemas/Formation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation non trouvée"
                    }
                }
            }
        },
        "/api/modifie_formation/{formation}": {
            "put": {
                "tags": [
                    "Formations"
                ],
                "summary": "Mettre à jour une formation",
                "description": "Met à jour les détails d'une formation spécifique.",
                "operationId": "870a0e12000a6022582d9b13b5869bb5",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID de la formation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_formation": {
                                        "type": "string"
                                    },
                                    "dure_formation": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation modifiée avec succès"
                    },
                    "404": {
                        "description": "Formation non modifiée"
                    }
                }
            }
        },
        "/api/supprimer_formation/{formation}": {
            "delete": {
                "tags": [
                    "Formations"
                ],
                "summary": "Supprimer une formation",
                "description": "Supprime une formation spécifique.",
                "operationId": "e2e6e5644cdabaf06b10eac185daba45",
                "parameters": [
                    {
                        "name": "formation",
                        "in": "path",
                        "description": "ID de la formation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation supprimée avec succès"
                    },
                    "404": {
                        "description": "Formation non supprimée"
                    }
                }
            }
        },
...
    },
...
}